services/iep-svc/requirements.txt
# IEP GraphQL Service Dependencies
# S1-11 Implementation - Python Package Requirements

# Core Framework
fastapi==0.104.1
strawberry-graphql[fastapi]==0.212.0
uvicorn[standard]==0.24.0

# Database & ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4

# CRDT & Collaboration
python-operational-transform==0.1.0
websockets==12.0
python-socketio==5.10.0

# Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# HTTP & Networking
httpx==0.25.2
aiohttp==3.9.1

# Testing Framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0

# Monitoring & Observability
prometheus-client==0.19.0
structlog==23.2.0

# File Upload & Storage
boto3==1.34.0
python-magic==0.4.27

# Cryptographic Operations
cryptography==41.0.8
hashlib-compat==1.0.1

# Environment & Configuration
python-dotenv==1.0.0
click==8.1.7

# JSON & Data Processing
orjson==3.9.10
jsonschema==4.20.0

# Background Tasks
celery[redis]==5.3.4
redis==5.0.1

# PDF & Document Processing
pypdf2==3.0.1
python-docx==1.1.0

# Email & Notifications
sendgrid==6.10.0
jinja2==3.1.2

# Utilities
python-slugify==8.0.1
shortuuid==1.0.11
humanize==4.9.0
